requirements.txt
flask
sentence-transformers
faiss-cpu
numpy
requests
python-dotenv